requirements.txt
streamlit
openai
opencv-python
moviepy
python-dotenv